chore: factored out autoPagination & maxPermissins for reuse
